Disentangle third party libraries (but keep support for now)
Add `which-key-inhibit-display-hook' to allow third-party libraries to prevent
which-key from displaying in certain circumstances. Add
`which-key-this-command-keys-function' as a customizable variable to allow
third-party libraries to teach which-key about custom symbols they insert into
the key sequence.

Add `which-key-evil-this-operator-p', `which-key-god-mode-self-insert-p' and
`which-key--god-mode-this-command-keys' to provide support for god-mode and evil
using the new mechanisms.