Use ;POS... position information in backtraces
Also code up adding this information for defvars, defconsts,
and cl-defmethods.  This is progress on bug #67455.

* lisp/emacs-lisp/byte-run.el (byte-run-position-vec): New
function.
(byte-run--fun-put-new-string): Tidy up the coding.
(byte-run-posify-existing-defaliases-1): Actually fset the
result to the pertinent symbol.
(byte-run-posify-existing-defvars/consts-1)
(byte-run-posify-existing-defvars/consts): New functions.

* lisp/emacs-lisp/bytecomp.el: Remove commented out old code.

* lisp/emacs-lisp/cl-generic.el (cl-defmethod): Insert the
;POS... structure into the form being defined, but using
explicit code rather than the (defining-symbol ...) declare
clause.  Add a byte-run-defined-form property to cl-defmethod
so the reader will position the generated symbols.

* lisp/emacs-lisp/cl-print.el (cl-print-object/cons): On
encountering a lambda or closure form, print the defining
symbol in braces.
(cl-print-object/compiled-function): Print the defining symbol
in braces.

* lisp/emacs-lisp/comp.el (comp-intern-func-in-ctxt): Pass
`also-pos' to `documentation' to get the doc's
;POS... information too.

* lisp/emacs-lisp/macroexp.el (macroexp--expand-all): Add code
to the defvar/defconst pcase arm to set `defining-symbol' to
the var/const being defined, and to strip the position from
that symbol.

* lisp/loadup.el: Set symbols-with-pos-enabled to t sooner.
Call the new function byte-run-posify-existing-defvars/consts
to posify variables/constants defined early in the boot
process.

* lisp/startup.el (normal-top-level): Set `debug' to debugger
for interactive sessions.

* src/data.c (syms_of_data): Give `defalias' a
byte-run-defined-form propery of 1, so that the reader will
position the symbol it defines.

* src/eval.c (Finternal__define_uninitialized_variable): Posify
the doc string argument, or give the symbol a
byte-run--early-defvar-const property early in the boot
procedure.
(handle_defvar_defconst_positions): New function.
(Fmacroexpand): Call the above function for a defvar/defconst.
(eval_sub): Also call the new function for a defvar/defconst.
(syms_of_eval): Declare two new symbols in byte-run.el.  Give
`defvar' and `defconst' byte-run-defined-form properties so as
to trigger the reader to position new symbols.