Fix 'srecode-all-template-hash' always returning nil
The issue was that the function always returned nil because
the closing parentheses were in the wrong position.  This
meant that the function couldn't give back the data it was
supposed to, making the whole srecode system not work
properly.
By moving the parentheses to the correct place, the function
now returns a hash table with all the templates it finds.
* lisp/cedet/srecode/find.el (srecode-all-template-hash):
Adjusted the position of closing parentheses so that the mhash
variable is returned correctly.  (Bug#70765)

Copyright-paperwork-exempt: yes